+++ src/github.rs
@@ -1,3 +1,9 @@
-pub(crate) mod client;
-pub(super) mod gql;
-pub(crate) mod pull_requests_summary;
+mod client;
+mod gql;
+mod pull_request;
+mod pull_requests;
+mod pull_requests_summary;
+
+pub(crate) use client::*;
+pub(crate) use pull_requests::*;
+pub(crate) use pull_requests_summary::*;
